Hide existing passes behind consistent pass APIs
The existing passes, including `ResourceAnalyser`, `FieldAnalyser`, and
`SemanticAnalyser` can have their implementation details hidden and
require callers to go through the front door.

Because not all passes are converted, this leads to some tests that have
multiple pass managers. Over time, these will merge into a single set of
passes and it will be more clear which tests require which passes.
