+++ src/ast/passes/portability_analyser.cpp
@@ -2,9 +2,29 @@
 
 #include <cstdlib>
 
+#include "ast/visitor.h"
 #include "types.h"
 
 namespace bpftrace::ast {
+
+namespace {
+
+// Checks if a script uses any non-portable bpftrace features that AOT
+// cannot handle.
+//
+// Over time, we expect to relax these restrictions as AOT supports more
+// features.
+class PortabilityAnalyser : public Visitor<PortabilityAnalyser> {
+public:
+  using Visitor<PortabilityAnalyser>::visit;
+  void visit(PositionalParameter &param);
+  void visit(Builtin &builtin);
+  void visit(Call &call);
+  void visit(Cast &cast);
+  void visit(AttachPoint &ap);
+};
+
+} // namespace
 
 void PortabilityAnalyser::visit(PositionalParameter &param)
 {
