+++ src/ast/passes/return_path_analyser.h
@@ -1,21 +1,9 @@
 #pragma once
 
 #include "ast/pass_manager.h"
-#include "ast/visitor.h"
 
 namespace bpftrace {
 namespace ast {
-
-class ReturnPathAnalyser : public Visitor<ReturnPathAnalyser, bool> {
-public:
-  // visit methods return true iff all return paths of the analyzed code
-  // (represented by the given node) return a value.
-  using Visitor<ReturnPathAnalyser, bool>::visit;
-  bool visit(Program &prog);
-  bool visit(Subprog &subprog);
-  bool visit(Jump &jump);
-  bool visit(If &if_stmt);
-};
 
 Pass CreateReturnPathPass();
