+++ tests/codegen/general.cpp
@@ -69,17 +69,12 @@ kprobe:f
   ClangParser clang;
   clang.parse(ast.root, *bpftrace);
 
-  driver.parse();
-  ASSERT_TRUE(ast.diagnostics().ok());
-
-  ast::SemanticAnalyser semantics(ast, *bpftrace);
-  semantics.analyse();
-  ASSERT_TRUE(ast.diagnostics().ok());
-
-  ast::ResourceAnalyser resource_analyser(*bpftrace);
-  resource_analyser.visit(ast.root);
-  bpftrace->resources = resource_analyser.resources();
-  ASSERT_TRUE(ast.diagnostics().ok());
+  auto ok = ast::PassManager()
+                .put(*bpftrace)
+                .add(ast::CreateSemanticPass())
+                .add(ast::CreateResourcePass())
+                .run(ast);
+  ASSERT_TRUE(ok && ast.diagnostics().ok());
 
   ast::CodegenLLVM codegen(ast, *bpftrace);
   codegen.generate_ir();
@@ -118,17 +113,19 @@ kprobe:f { 1; } kprobe:d { 1; }
 )");
   MockBPFtrace bpftrace;
   EXPECT_CALL(bpftrace, add_probe(_, _, _, _)).Times(2);
+  // Override to mockbpffeature.
+  bpftrace.feature_ = std::make_unique<MockBPFfeature>(true);
 
   Driver driver(ast, bpftrace);
 
   driver.parse();
   ASSERT_TRUE(ast.diagnostics().ok());
 
-  // Override to mockbpffeature.
-  bpftrace.feature_ = std::make_unique<MockBPFfeature>(true);
-  ast::SemanticAnalyser semantics(ast, bpftrace);
-  semantics.analyse();
-  ASSERT_TRUE(ast.diagnostics().ok());
+  auto ok = ast::PassManager()
+                .put<BPFtrace>(bpftrace)
+                .add(ast::CreateSemanticPass())
+                .run(ast);
+  ASSERT_TRUE(ok && ast.diagnostics().ok());
 
   ast::CodegenLLVM codegen(ast, bpftrace);
   codegen.generate_ir();
